Workaround RT#66040: force Locale::Maketext upgrade
RT#66040 is not a bug in Parse::nm.
Instead it is a difference of behavior in perl 5.12 that makes
Locale::Maketext < 1.16 print warnings. As we are using Test::NoWarnings
in tests as failure and Locale::Maketext is used indirectly (via perl
open '|-' which seems to use IPC::Cmd), we are spammed with CPAN Testers
reports of users which have an old Locale::Maketext.
This patch forces a dependency on a newer Locale::Maketext if it is too
old.